[#6] Refactor: modal Trigger 구조 변경
